Move ColorModeSwitcher to components folder
