+++ frontend/src/components/Header.tsx
@@ -18,7 +18,7 @@ import {
 import React, { useState } from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { Link } from "react-router-dom";
-import { ColorModeSwitcher } from "../ColorModeSwitcher";
+import { ColorModeSwitcher } from "./ColorModeSwitcher";
 import { logout, auth } from "../firebase";
 import { MdEmail } from "react-icons/all";
 import { User } from "firebase/auth";
